documented code, changed api route names to match standard convention, documented api routes on routes.md file
