+++ main/src/app/api/get-available-session/route.tsx
@@ -3,33 +3,59 @@ import dbConnect from "@/lib/db";
 import Conversation from "@/models/Conversation";
 import { UserDocument } from "@/datatypes/dataTypes";
 
+/**
+ * Handles GET requests to fetch the next available session ID for a user.
+ *
+ * Input Format (Query Parameters):
+ * - user_id (string, required): The ID of the user for whom the next session ID is being requested.
+ * 
+ * Output Format (Response):
+ * - Success Response (200):
+ *   {
+ *     "next_session_id": "1" (or another ID if the user has sessions)
+ *   }
+ * - Error Responses:
+ *   - 400: Missing user_id
+ *     {
+ *       "error": "Missing user_id"
+ *     }
+ *   - 500: Internal server error
+ *     {
+ *       "error": "Internal Server Error"
+ *     }
+ *
+ * @param req - The incoming Next.js request object containing query parameters.
+ * @returns JSON response with the next session ID or error message.
+ */
 export async function GET(req: NextRequest): Promise<NextResponse> {
-  // GET request check
+  
+  // Ensure the request method is GET, otherwise return a 405 (Method Not Allowed) error.
   if (req.method !== "GET") {
     return NextResponse.json(
-      { error: "Only POST requests are allowed" },
+      { error: "Only GET requests are allowed" },
       { status: 405 }
     );
   }
+
   try {
-    // Extract query parameters from the URL
+    // Extract query parameters, specifically user_id
     const { searchParams } = new URL(req.url);
     const user_id = searchParams.get("user_id");
 
-    // Validate that user_id is provided
+    // If user_id is missing in the query parameters, return a 400 (Bad Request) error
     if (!user_id) {
       return NextResponse.json({ error: "Missing user_id" }, { status: 400 });
     }
 
-    // Connect to MongoDB using Mongoose
+    // Connect to the database
     await dbConnect();
 
-    // Retrieve the user document using the Conversation model
+    // Find the user document by user_id
     const userDocument = (await Conversation.findOne({
       _id: user_id,
     }).lean()) as UserDocument;
 
-    // If user doesn't exist or doesn't have any sessions, return session_id as 1
+    // If no sessions exist for the user, return next_session_id as "1"
     if (
       !userDocument ||
       !userDocument.sessions ||
@@ -38,11 +64,11 @@ export async function GET(req: NextRequest): Promise<NextResponse> {
       return NextResponse.json({ next_session_id: 1 }, { status: 200 });
     }
 
-    // Convert the Mongoose document to a plain JS object and retrieve the keys of the sessions object
+    // Get the list of existing session IDs
     const sessionIds = Object.keys(userDocument.sessions);
 
+    // Determine the next available session ID by finding the smallest unused session ID
     let nextSessionId = 1;
-    // Find the first missing session_id starting from 1
     while (sessionIds.includes(nextSessionId.toString())) {
       nextSessionId++;
     }
@@ -53,6 +79,7 @@ export async function GET(req: NextRequest): Promise<NextResponse> {
       { status: 200 }
     );
   } catch (error) {
+    // Catch and handle any errors, returning a 500 (Internal Server Error)
     console.error("Error fetching next session ID:", error);
     return NextResponse.json(
       { error: "Internal Server Error" },
